Optimize TracingProfiler and RecordingStatistics
`TracingProfiler._profile` calculates the time at the initial line for
leaving record.  Then calculates again for entering record.  Timer is
expensive.  Now it calculates the time just once.

The lock of `RecordingStatistics` causes a low performance.  But it seems to be
not required.  In my test, the callback of `threading.setprofile` was called
atomically.  I removed the lock.